+++ mahran/src/App.js
@@ -1,25 +1,12 @@
-import React, { useContext } from "react";
-import Navbar from "./Day4/components/Navbar";
-import Login from "./Day4/components/Login";
-import { AuthProvider, AuthContext } from "./Day4/AuthContext";
+import React from "react";
+import HomeWithStore from "./Day4/Home";
 
 function App() {
-  const authContext = useContext(AuthContext);
-
   return (
-    <div className="container">
-      <Navbar />
-      {authContext.auth.email ? "welcome" : <Login />}
+    <div className="App">
+      <HomeWithStore />
     </div>
   );
 }
 
-function AppWithStore() {
-  return (
-    <AuthProvider>
-      <App />
-    </AuthProvider>
-  );
-}
-
-export default AppWithStore;
+export default App;
